+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 ### __Breaking Changes__
 
 ### Added
+- LayerTree: added a new parameter at layer `filterRefId` as an Index of a configured filter Layer.
 
 ### Changed